Add cache middleware
